projects/telegram_echo_bot/telegram_bot_simple.py
# relationship_bot.py
from typing import Dict, List, Optional, AsyncGenerator
# from cosette import Chat, models
# import cosette
import time
from datetime import datetime, timedelta
import random
# from ..utils.localization import get_text
# from ..database.db_manager import DatabaseManager
# from ..utils.logger import setup_logger
# from ..utils.validation import validate_string, validate_user_input, validate_user_id
# from ..utils.decorators import admin_required
# from ..utils.custom_exceptions import ValidationError, DatabaseException, AIException
# from ..utils.referral import (
#     generate_referral_code,
#     handle_referral,
#     apply_referral,
#     get_referrer_id,
#     add_referral_bonus,
#     get_or_create_referral_code,
#     get_user_referral_data,
#     user_has_referral,
# )
# from ..config.settings import config, REFERRAL_BONUS_AMOUNT
import asyncio
import logging
import os
from telegram import Update
from telegram.ext import Application, CommandHandler, MessageHandler, filters, ContextTypes

# Enable logging
logging.basicConfig(
    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s", level=logging.INFO
)
logger = logging.getLogger(__name__)

# Simulated database for demonstration purposes
class SimpleDatabaseManager:
    def __init__(self):
        self.users = {}
        self.sessions = {}
        self.interactions = []

    async def create_or_get_active_session(self, user_id):
        if user_id not in self.sessions:
            self.sessions[user_id] = {"id": len(self.sessions) + 1}
        return self.sessions[user_id]

    async def log_interaction(self, session_id, user_id, interaction_type, content):
        self.interactions.append({
            "session_id": session_id,
            "user_id": user_id,
            "type": interaction_type,
            "content": content,
            "timestamp": datetime.now()
        })

    async def get_session_interaction_count(self, session_id):
        return len([i for i in self.interactions if i["session_id"] == session_id])

    async def update_session_activity(self, session_id):
        pass  # Simplified for this example

class RelationshipBot:
    def __init__(self, db_manager: SimpleDatabaseManager, bot_username: str):
        self.db_manager = db_manager
        self.bot_username = bot_username
        self.feedback_threshold = 5
        self.feedback_state = {}
        self.logger = logging.getLogger(__name__)
        # self.chat_instances = {
        #     "gpt-4o-mini": Chat(model="gpt-4o-mini-2024-07-18"),
        #     "gpt-4o": Chat(model="gpt-4o-2024-08-06"),
        # }

    def get_text(self, key, **kwargs):
        # Simplified localization function
        texts = {
            "contact_info": "Contact us at: support@example.com",
            "feedback_request": "How was your experience? Please rate with 👍 or 👎.",
            "error_handling_message": "Sorry, there was an error processing your message. Please try again.",
            "construct_prompt": "User age: {age}, gender: {gender}. Partner: {partner_name}. Message: {message}",
            "feedback_thanks_positive": "Thank you for your positive feedback!",
            "feedback_request_details": "Could you please provide more details about your experience?",
            "feedback_invalid_response": "Please respond with 👍 or 👎.",
            "feedback_thanks_detailed": "Thank you for your detailed feedback. We appreciate your input!"
        }
        return texts.get(key, "").format(**kwargs)

    async def handle_message(self, user_id: int, message: str) -> AsyncGenerator[str, None]:
        try:
            session = await self.db_manager.create_or_get_active_session(user_id)
            await self.db_manager.log_interaction(session["id"], user_id, "user_message", message)

            if user_id in self.feedback_state:
                async for response in self.handle_feedback_response(user_id, session["id"], message):
                    yield response
                return

            response = f"Echo: {message}"  # Simplified response generation
            yield response

            interaction_count = await self.db_manager.get_session_interaction_count(session["id"])
            if interaction_count % self.feedback_threshold == 0:
                yield self.get_text("feedback_request")
                self.feedback_state[user_id] = {
                    "state": "awaiting_feedback",
                    "interaction_id": None,
                }

            await self.db_manager.update_session_activity(session["id"])

        except Exception as e:
            self.logger.error(f"Unexpected error {user_id}: {str(e)}")
            yield self.get_text("error_handling_message")

    async def handle_feedback_response(self, user_id: int, session_id: int, message: str) -> AsyncGenerator[str, None]:
        state = self.feedback_state.get(user_id, {}).get("state")
        interaction_id = self.feedback_state.get(user_id, {}).get("interaction_id")

        if not state or not interaction_id:
            yield self.get_text("error_handling_message")
            return

        if state == "awaiting_feedback":
            if message in ["👍", "👎"]:
                is_positive = message == "👍"
                if is_positive:
                    del self.feedback_state[user_id]
                    yield self.get_text("feedback_thanks_positive")
                else:
                    self.feedback_state[user_id]["state"] = "awaiting_text_feedback"
                    yield self.get_text("feedback_request_details")
            else:
                yield self.get_text("feedback_invalid_response")
        elif state == "awaiting_text_feedback":
            del self.feedback_state[user_id]
            yield self.get_text("feedback_thanks_detailed")

# Telegram Bot integration
async def start(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
    await update.message.reply_text('Welcome to the RelationshipBot! Send me a message and I will respond.')

async def handle_message(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
    user_id = update.effective_user.id
    message = update.message.text
    bot = context.bot_data['relationship_bot']
    
    async for response in bot.handle_message(user_id, message):
        await update.message.reply_text(response)

def main() -> None:
    token = os.getenv("TELEGRAM_BOT_TOKEN")
    if not token:
        raise RuntimeError("Missing TELEGRAM_BOT_TOKEN environment variable")
    application = Application.builder().token(token).build()

    # Create RelationshipBot instance
    db_manager = SimpleDatabaseManager()
    relationship_bot = RelationshipBot(db_manager, "Soul_Matchbot")
    
    # Store RelationshipBot instance in bot_data
    application.bot_data['relationship_bot'] = relationship_bot

    # Add handlers
    application.add_handler(CommandHandler("start", start))
    application.add_handler(MessageHandler(filters.TEXT & ~filters.COMMAND, handle_message))

    # Run the bot until the user presses Ctrl-C
    application.run_polling()

if __name__ == "__main__":
    main()
