projects/telegram_echo_bot/requirements.txt
python-telegram-bot>=20,<21
fastapi>=0.110
modal
requests>=2.31
